Create Initial Coding
Created Python coding to calculate gross pay from user input hours and rate (using the input function). Used the float function to convert the user input from strings to float so they can be used in calculations. Used 'if' and 'else' to determine if overtime calculations are needed and to run the correct equation to calculate pay (with or without overtime).